Rename cgp_macro_lib::getter_component to derive_getter
