+++ crates/cgp-macro-lib/src/derive_getter/symbol.rs
