+++ crates/cgp-macro-lib/src/derive_getter/use_field.rs
@@ -3,8 +3,8 @@ use syn::punctuated::Punctuated;
 use syn::token::Plus;
 use syn::{parse2, Generics, ItemImpl, ItemTrait, TypeParamBound};
 
-use crate::getter_component::getter_field::GetterField;
-use crate::getter_component::{derive_getter_constraint, derive_getter_method, ContextArg};
+use crate::derive_getter::getter_field::GetterField;
+use crate::derive_getter::{derive_getter_constraint, derive_getter_method, ContextArg};
 use crate::parse::ComponentSpec;
 
 pub fn derive_use_field_impl(
