+++ crates/cgp-macro-lib/src/derive_getter/use_fields.rs
@@ -6,9 +6,9 @@ use syn::punctuated::Punctuated;
 use syn::token::Plus;
 use syn::{parse2, ItemImpl, ItemTrait, TypeParamBound};
 
-use crate::getter_component::getter_field::GetterField;
-use crate::getter_component::symbol::symbol_from_string;
-use crate::getter_component::{derive_getter_constraint, derive_getter_method, ContextArg};
+use crate::derive_getter::getter_field::GetterField;
+use crate::derive_getter::symbol::symbol_from_string;
+use crate::derive_getter::{derive_getter_constraint, derive_getter_method, ContextArg};
 use crate::parse::ComponentSpec;
 
 pub fn derive_use_fields_impl(
