+++ crates/cgp-macro-lib/src/derive_getter/with_provider.rs
@@ -2,8 +2,8 @@ use proc_macro2::Span;
 use quote::{quote, ToTokens};
 use syn::{parse2, Generics, Ident, ItemImpl, ItemTrait};
 
-use crate::getter_component::getter_field::GetterField;
-use crate::getter_component::{derive_getter_method, ContextArg};
+use crate::derive_getter::getter_field::GetterField;
+use crate::derive_getter::{derive_getter_method, ContextArg};
 use crate::parse::ComponentSpec;
 
 pub fn derive_with_provider_impl(
