+++ crates/cgp-macro-lib/src/entrypoints/cgp_auto_getter.rs
@@ -2,7 +2,7 @@ use proc_macro2::{Span, TokenStream};
 use quote::quote;
 use syn::{Error, Ident, ItemTrait};
 
-use crate::getter_component::{derive_blanket_impl, parse_getter_fields};
+use crate::derive_getter::{derive_blanket_impl, parse_getter_fields};
 
 pub fn cgp_auto_getter(attr: TokenStream, body: TokenStream) -> syn::Result<TokenStream> {
     if !attr.is_empty() {
