+++ crates/cgp-macro-lib/src/lib.rs
@@ -14,10 +14,10 @@ pub(crate) mod check_components;
 pub(crate) mod delegate_components;
 pub(crate) mod derive_component;
 pub(crate) mod derive_context;
+pub(crate) mod derive_getter;
 pub(crate) mod derive_provider;
 mod entrypoints;
 pub(crate) mod for_each_replace;
-pub(crate) mod getter_component;
 pub(crate) mod parse;
 pub(crate) mod preset;
 pub(crate) mod type_component;
